Reverted last commit

+++ src/main/java/nl/mvdr/adventofcode/adventofcode2022/day11/Item.java
@@ -1,13 +1,11 @@
 package nl.mvdr.adventofcode.adventofcode2022.day11;
 
-import java.math.BigInteger;
-
 /**
  * Representation of an item which has been stolen by a monkey.
  *
  * @author Martijn van de Rijdt
  */
-record Item(BigInteger worryLevel) {
+record Item(int worryLevel) {
     
     @Override
     public String toString() {
